# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BetPredict Pro API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface TeamPrediction {
  winProbability: number;
  form: string;
  recentGoalsScored: number;
  recentGoalsConceded: number;
  homeAwayRecord: string;
}

export type MatchStatus = (typeof MatchStatus)[keyof typeof MatchStatus];

export const MatchStatus = {
  upcoming: "upcoming",
  live: "live",
  finished: "finished",
} as const;

export type MatchPrediction =
  (typeof MatchPrediction)[keyof typeof MatchPrediction];

export const MatchPrediction = {
  home: "home",
  draw: "draw",
  away: "away",
  home_or_draw: "home_or_draw",
  away_or_draw: "away_or_draw",
  btts: "btts",
  over25: "over25",
} as const;

export interface Match {
  id: string;
  homeTeam: string;
  awayTeam: string;
  homeTeamLogo?: string;
  awayTeamLogo?: string;
  league: string;
  leagueId: string;
  country: string;
  countryFlag?: string;
  matchDate: string;
  matchTime: string;
  status: MatchStatus;
  homeScore?: number | null;
  awayScore?: number | null;
  homeWinOdds: number;
  drawOdds: number;
  awayWinOdds: number;
  confidenceScore: number;
  prediction: MatchPrediction;
  predictionLabel: string;
  valueRating: number;
  isHot: boolean;
}

export interface H2HRecord {
  date: string;
  homeTeam: string;
  awayTeam: string;
  homeScore: number;
  awayScore: number;
}

export type MatchDetail = Match & {
  homeTeamPrediction: TeamPrediction;
  awayTeamPrediction: TeamPrediction;
  headToHead: H2HRecord[];
  analysisNotes: string[];
  bttsProb: number;
  over25Prob: number;
  expectedHomeGoals: number;
  expectedAwayGoals: number;
};

export interface BetOfTheDaySelection {
  matchId: string;
  homeTeam: string;
  awayTeam: string;
  league: string;
  country: string;
  prediction: string;
  predictionLabel: string;
  odds: number;
  confidenceScore: number;
  matchDate: string;
  matchTime: string;
}

export interface BetOfTheDay {
  id: string;
  generatedAt: string;
  totalOdds: number;
  averageConfidence: number;
  selections: BetOfTheDaySelection[];
  analysisNote: string;
}

export interface League {
  id: string;
  name: string;
  country: string;
  countryFlag?: string;
  logo?: string;
  matchCount: number;
}

export interface Country {
  name: string;
  flag?: string;
  leagueCount: number;
  matchCount: number;
}

export interface MatchListResponse {
  matches: Match[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export interface StatsSummary {
  totalMatches: number;
  hotGames: number;
  avgConfidence: number;
  successRate: number;
  totalLeagues: number;
  totalCountries: number;
  todayMatches: number;
}

export interface LeagueAccuracy {
  league: string;
  country: string;
  accuracy: number;
  totalPredictions: number;
  correct: number;
}

export type ListMatchesParams = {
  search?: string;
  leagueId?: string;
  country?: string;
  date?: string;
  status?: ListMatchesStatus;
  page?: number;
  limit?: number;
};

export type ListMatchesStatus =
  (typeof ListMatchesStatus)[keyof typeof ListMatchesStatus];

export const ListMatchesStatus = {
  upcoming: "upcoming",
  live: "live",
  finished: "finished",
} as const;

export type GetHotGamesParams = {
  limit?: number;
};

export type ListLeaguesParams = {
  country?: string;
};