# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: BetPredict Pro API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /matches:
    get:
      operationId: listMatches
      tags: [matches]
      summary: List matches with filters
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: leagueId
          in: query
          schema:
            type: string
        - name: country
          in: query
          schema:
            type: string
        - name: date
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [upcoming, live, finished]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of matches
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchListResponse"
  /matches/{id}:
    get:
      operationId: getMatch
      tags: [matches]
      summary: Get a match with full prediction details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Match with predictions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /hot-games:
    get:
      operationId: getHotGames
      tags: [matches]
      summary: Get high-value hot games
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Hot games list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Match"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    TeamPrediction:
      type: object
      properties:
        winProbability:
          type: number
        form:
          type: string
        recentGoalsScored:
          type: number
        recentGoalsConceded:
          type: number
        homeAwayRecord:
          type: string
      required:
        - winProbability
        - form
        - recentGoalsScored
        - recentGoalsConceded
        - homeAwayRecord
    Match:
      type: object
      properties:
        id:
          type: string
        homeTeam:
          type: string
        awayTeam:
          type: string
        homeTeamLogo:
          type: string
        awayTeamLogo:
          type: string
        league:
          type: string
        leagueId:
          type: string
        country:
          type: string
        countryFlag:
          type: string
        matchDate:
          type: string
        matchTime:
          type: string
        status:
          type: string
          enum: [upcoming, live, finished]
        homeScore:
          type: integer
          nullable: true
        awayScore:
          type: integer
          nullable: true
        homeWinOdds:
          type: number
        drawOdds:
          type: number
        awayWinOdds:
          type: number
        confidenceScore:
          type: number
        prediction:
          type: string
          enum: [home, draw, away, home_or_draw, away_or_draw, btts, over25]
        predictionLabel:
          type: string
        valueRating:
          type: number
        isHot:
          type: boolean
      required:
        - id
        - homeTeam
        - awayTeam
        - league
        - leagueId
        - country
        - matchDate
        - matchTime
        - status
        - homeWinOdds
        - drawOdds
        - awayWinOdds
        - confidenceScore
        - prediction
        - predictionLabel
        - valueRating
        - isHot
    MatchDetail:
      allOf:
        - $ref: "#/components/schemas/Match"
        - type: object
          properties:
            homeTeamPrediction:
              $ref: "#/components/schemas/TeamPrediction"
            awayTeamPrediction:
              $ref: "#/components/schemas/TeamPrediction"
            headToHead:
              type: array
              items:
                $ref: "#/components/schemas/H2HRecord"
            analysisNotes:
              type: array
              items:
                type: string
            bttsProb:
              type: number
            over25Prob:
              type: number
            expectedHomeGoals:
              type: number
            expectedAwayGoals:
              type: number
          required:
            - homeTeamPrediction
            - awayTeamPrediction
            - headToHead
            - analysisNotes
            - bttsProb
            - over25Prob
            - expectedHomeGoals
            - expectedAwayGoals
    H2HRecord:
      type: object
      properties:
        date:
          type: string
        homeTeam:
          type: string
        awayTeam:
          type: string
        homeScore:
          type: integer
        awayScore:
          type: integer
      required: [date, homeTeam, awayTeam, homeScore, awayScore]
...
    MatchListResponse:
      type: object
      properties:
        matches:
          type: array
          items:
            $ref: "#/components/schemas/Match"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required: [matches, total, page, limit, totalPages]
...