# lib/api-zod/src/generated/types/h2HRecord.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BetPredict Pro API
 * OpenAPI spec version: 0.1.0
 */

export interface H2HRecord {
  date: string;
  homeTeam: string;
  awayTeam: string;
  homeScore: number;
  awayScore: number;
}