# lib/api-zod/src/generated/types/match.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BetPredict Pro API
 * OpenAPI spec version: 0.1.0
 */
import type { MatchPrediction } from "./matchPrediction";
import type { MatchStatus } from "./matchStatus";

export interface Match {
  id: string;
  homeTeam: string;
  awayTeam: string;
  homeTeamLogo?: string;
  awayTeamLogo?: string;
  league: string;
  leagueId: string;
  country: string;
  countryFlag?: string;
  matchDate: string;
  matchTime: string;
  status: MatchStatus;
  homeScore?: number | null;
  awayScore?: number | null;
  homeWinOdds: number;
  drawOdds: number;
  awayWinOdds: number;
  confidenceScore: number;
  prediction: MatchPrediction;
  predictionLabel: string;
  valueRating: number;
  isHot: boolean;
}