# lib/api-zod/src/generated/types/matchDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BetPredict Pro API
 * OpenAPI spec version: 0.1.0
 */
import type { H2HRecord } from "./h2HRecord";
import type { Match } from "./match";
import type { TeamPrediction } from "./teamPrediction";

export type MatchDetail = Match & {
  homeTeamPrediction: TeamPrediction;
  awayTeamPrediction: TeamPrediction;
  headToHead: H2HRecord[];
  analysisNotes: string[];
  bttsProb: number;
  over25Prob: number;
  expectedHomeGoals: number;
  expectedAwayGoals: number;
};