# lib/api-zod/src/generated/types/matchListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BetPredict Pro API
 * OpenAPI spec version: 0.1.0
 */
import type { Match } from "./match";

export interface MatchListResponse {
  matches: Match[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}