# lib/api-zod/src/generated/types/teamPrediction.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BetPredict Pro API
 * OpenAPI spec version: 0.1.0
 */

export interface TeamPrediction {
  winProbability: number;
  form: string;
  recentGoalsScored: number;
  recentGoalsConceded: number;
  homeAwayRecord: string;
}